+++ Changelog.md
@@ -78,6 +78,9 @@
 
 1. Removes one function call and replaces a manual swap with std::swap in ODE heightfield.
     * [Pull request #2114](https://bitbucket.org/osrf/gazebo/pull-request/2114)
+    
+1. New world event: BeforePhysicsUpdate
+    * [Issue #1851](https://bitbucket.org/osrf/gazebo/issues/1851)
 
 1. Model editor updates
     1. Fix setting relative pose after alignment during joint creation.